webapp/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
jinja2>=3.1.0
httpx>=0.26.0
anthropic>=0.18.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
python-frontmatter>=1.1.0
markdown-it-py>=3.0.0
typst>=0.11.0
pillow>=10.2.0
